+++ inst/IAOReader/server.R
@@ -13,8 +13,7 @@ server <- function(input, output, session) {
     plot_settings(input, output, session)
     input_settings(input, output, session)
 
-    # TODO: update with real statistics.
-    output[["summary_table"]] <- renderTable({
-        data.frame("X" = 1:3, "Y" = 4:6, "Z" = 7:9, "XYZ" = 11:13)
-    })
+    observeEvent(input[["download_svg"]], {
+        session$sendCustomMessage("download_svg", 1)
+    }, ignoreInit = TRUE)
 }