+++ inst/IAOReader/www/js/iaoreader.js
@@ -231,6 +231,16 @@ let IAOReader = class {
         return results;
     }
 
+    calculate_summary_table() {
+        var lambda_values = {};
+
+        for (var i = this.x_min; i <= this.x_max; i++) {
+            lambda_values[i] = this.lambda(i);
+        }
+
+        Shiny.setInputValue("summary_table", lambda_values);
+    }
+
 
     /* -------------------------------------------------------------------------
      * Getters
@@ -347,11 +357,9 @@ let IAOReader = class {
     }
 
     set title_font_size(size) {
-        // This makes the tile centered at the initial y.
-        if (10 <= size && size <= 72) {
-            this.title.attr("y", this.margin.top * 0.60 + (size - 20) / 2 + "px");
-            this.title.attr("font-size", size + "px");
-        }
+        // This makes the tilte centered at the initial y.
+        this.title.attr("y", this.margin.top * 0.60 + (size - 20) / 2 + "px");
+        this.title.attr("font-size", size + "px");
     }
 
     set title_color(color) {
@@ -390,6 +398,8 @@ let IAOReader = class {
 
         this.unmark_lines(this.vert_click_mark);
         this.vert_click.style("visibility", "hidden");
+
+        this.calculate_summary_table();
     }
 
     draw_x_axis() {
@@ -398,6 +408,7 @@ let IAOReader = class {
 
     draw_y_axis() {
         this.y_axis.call(d3.axisLeft().scale(this.y_scale));
+        this.y_axis.selectAll("g.tick").remove();
     }
 
     draw_lines() {
@@ -442,7 +453,7 @@ let IAOReader = class {
         // Adding new values to the vert.
         for (const [file_name, y] of Object.entries(heights)) {
             var lambda_val = Math.round(lambda_values[file_name] * 100);
-            var dx = 13 + 4 * lambda_val.toString().length + horizontal_padding;
+            var x = 13 + 4 * lambda_val.toString().length + horizontal_padding;
 
             var rect = vert.append("rect")
                 .attr("class", "lambda")
@@ -451,10 +462,9 @@ let IAOReader = class {
 
             var text = vert.append("text")
                 .attr("class", "lambda")
-                .attr("y", this.y_scale(y))
+                .attr("x", top_placement ? -x : x)
+                .attr("y", this.y_scale(y) + (top_placement ? -11 : 20))
                 .attr("fill", this.file_color(file_name))
-                .attr("dx", top_placement ? -dx : dx)
-                .attr("dy", top_placement ? -11 : 20)
                 .text(lambda_val + "%");
 
             this.draw_text_bbox(text, rect);
@@ -463,8 +473,8 @@ let IAOReader = class {
             // in order to not obstruct axis label.
             var rect_lower_limit = +rect.attr("y") + +rect.attr("height");
             if (this.y_scale.invert(rect_lower_limit) < 0) {
-                var dx_adj = vert.select("rect.axis-label").attr("width") / 2;
-                text.attr("dx", dx + (top_placement ? -dx_adj : dx_adj));
+                var x_adj = vert.select("rect.axis-label").attr("width") / 2;
+                text.attr("x", x + (top_placement ? -x_adj : x_adj));
                 this.draw_text_bbox(text, rect);
             }
         }
@@ -564,6 +574,15 @@ let IAOReader = class {
 
         // This moves vert to it's current position to redraw lambda values.
         this.move_vert(vert, this.x_scale.invert(vert_px_position));
+    }
+
+
+    /* -------------------------------------------------------------------------
+     * Downloading plot as SVG
+     * ---------------------------------------------------------------------- */
+
+    download_svg() {
+        download_svg_node(this.svg.node());
     }
 }